requirements.txt
customtkinter
opencv-python
Pillow
ultralytics
zeep